Refs #228: Setup Cakefile for builds and setup pre/post publish tasks.
With this setup, the package published to the NPM repo will not list
`browserify` or `coffeeify` as dependencies. You can still install
direct from a git repo. In that case it will need browserify and
coffeeify.

I also cleaned it up a bit so it only builds the 3 javascript files that
are actually used.