Fix implicit parens edge-case (for last line)
This is an attempt at fixing #302. Basically if you have strict set to
false as part of the no_implicit_parens config, it should allow implicit
parens when it's done across multiple lines.

Before the change was done by detecting for a newline. Unfortunately it
seems that CoffeeScript does not add a newline property so you can have
implicit parens on the last line of code.