Fixes method chaining looking out of scope
This PR is a rewrite of the original code that was used to generate the
correct INDENT value for INDENT tokens when a user wrote code that used
method chaining with each new method on it's own particular line.

There is less redundant code, with a new function called
`getCorrectIndent` that will return the corrected INDENT value that
should be there instead of what CoffeeScript original generates.

This also fixes #305 where the original `isChainedCall` function was
looking out of its scope when checking for proper indenting in the
current method chain.