New rule: Ensure List Comprehensions
This new rule makes sure that the user remembers to put parentheses
around any list comprehensions they make. Not putting parentheses around
a list comprehension is generally a mistake since they actually want the
whole list and not just a value.

This rule will ignore for-loops with a block, and will ignore for-loops
where an identifier on the left-hand-side of the assignment operator is
either a key, value or index identifier of the for loop

E.g.
```coffeescript
newConfig[key] = value for key, value of config
```