Reworked no_implicit_braces for non-standard cases
This PR reworks no_implicit_braces to better handle edge cases

CoffeeScript generates implicit braces starting at the first instance
method of the class (aka, the prototypes of the Object), it is actually
not generating it for the Class persay. If there is code that is not
part of the prototype/instance methods of the class (akin to the class
methods/properties of the Object), then it throws off where the '{' is
generated.

Fixes #458, Fixes #459