Ignore '=' in non-comprehension nested-for-loops
This fixes the linter seeing an equal sign inside of a nested loop.

Normally the linter checks for an '=' to determine if the for loop is a
possible comprehension. This also picked up an equal sign inside of a
nested loop.

This fixes the linter seeing an euqal sign inside of a nested loop by
checking the number of parens ('(' or ')') and only triggering the rule
if there is a balanced number of parens.

Fixes #486