coffeelint: inline rules now display error messages
There was a bug where if a rule that was not in your config was
"enabled" via an inline comment, it would not have a display message set
if the rule was broken.

Also gave me some time to cleanup the rule setting code, which was using
extra for loops for some reason.