colon_assignment_spacing: add opt. properties min_left, min_right
Usage:
    spacing:
        min_left: 1
        min_right: 2

This means that the space to the left of the assignment-colon has to be at
least 1 whitespace; and the space to the right of the assignment-colon has to
be at least 2 whitespaces.

The `left` and `right` are ignored if `min_left` and `min_right` are
non-negative. By default, `min_left` and `min_right` are ignored as they are
set to `-1` by default.

Also, the code of the rule was simplified. And finally, a test case was added
as well.